Read SQL Server's information schema at the widths it declares
Every query against SQL Server failed at the first row of metadata with
InvalidCastException: Unable to cast object of type 'System.Byte' to type
'System.Int32'. INFORMATION_SCHEMA.COLUMNS types NUMERIC_PRECISION as a
tinyint, SqlClient carries that onto the DataTable it builds, and
DataRow.Field<int?> unboxes rather than converts. Any table with a numeric
column threw, whatever the query projected, because the row type is derived
from every column of the table.

Reading those counts through Convert.ToInt32 is the fix. The report's own
guess -- AdoReaderUtil.GetInt -- was not it; the failure was in metadata,
before a row was read.

Making a SQL Server table readable took four more things, each found by a
test that could not have been written against SQLite:

  - ParseDbType knew eight type names out of the server's thirty and threw
    NotImplementedException on the rest, so a bit, a decimal or a date made
    the whole table unreadable. It now names every one, and an unknown name
    falls to DbType.Object, which AdoTable already maps to OTHER.

  - DbType.Byte is unsigned 0..255 and Calcite's TINYINT is signed, so a
    tinyint of 200 came back as -56. SMALLINT is the narrowest type holding
    it.

  - The typed accessors on DbDataReader cast rather than convert, so a
    column narrower than the Calcite type Calcite chose for it throws --
    a tinyint under a SMALLINT, a Guid under a CHAR(36). They convert now.

  - A TIMESTAMP was cast through DateTimeOffset, which reads an unspecified
    DateTime.Kind as local time: every timestamp was out by the machine's
    offset, the inverse of what ParameterBinder.ConvertTimestamp writes and
    CalciteResultValue decodes. GetDate already avoided this hazard and
    said so in a comment. A TIMESTAMP_TZ went through GetDateTime, which
    SqlClient refuses outright for a datetimeoffset.

SqlServerFixture creates a LocalDB database and drops it with the test;
SqlServerQueryTests and SqlServerReaderUtilTests skip where no instance is
reachable, which is every platform but Windows and a Windows machine
without LocalDB. 61 tests, and the suite passes with them skipped.

Closes #24

+++ src/Apache.Calcite.Adapter.AdoNet/AdoTable.cs
@@ -161,8 +161,11 @@ namespace Apache.Calcite.Adapter.AdoNet
                     return typeFactory.createSqlType(SqlTypeName.VARCHAR, size);
                 case DbType.Binary:
                     return typeFactory.createSqlType(SqlTypeName.VARBINARY, size);
+                // DbType.Byte is the unsigned 0..255 one and TINYINT is signed, so the top half of its range
+                // comes back negative: SQL Server's tinyint 200 read as a TINYINT is -56. SMALLINT is the
+                // narrowest Calcite type that holds it
                 case DbType.Byte:
-                    return typeFactory.createSqlType(SqlTypeName.TINYINT);
+                    return typeFactory.createSqlType(SqlTypeName.SMALLINT);
                 case DbType.Boolean:
                     return typeFactory.createSqlType(SqlTypeName.BOOLEAN);
                 // the scale money carries in every provider that has a distinct type for it
